+++ src/migrations/Install.php
@@ -5,6 +5,7 @@ namespace ether\seo\migrations;
 use craft\db\Migration;
 use ether\seo\records\RedirectRecord;
 use ether\seo\records\SitemapRecord;
+use ether\seo\Seo;
 
 class Install extends Migration
 {
@@ -44,9 +45,10 @@ class Install extends Migration
 			[
 				'id' => $this->primaryKey(),
 
-				'uri'  => $this->string(255)->notNull(),
-				'to'   => $this->string(255)->notNull(),
-				'type' => $this->enum('type', ['301', '302'])->notNull(),
+				'uri'    => $this->string(255)->notNull(),
+				'to'     => $this->string(255)->notNull(),
+				'type'   => $this->enum('type', ['301', '302'])->notNull(),
+				'siteId' => $this->integer()->null(),
 
 				'dateCreated' => $this->dateTime()->notNull(),
 				'dateUpdated' => $this->dateTime()->notNull(),